refactor(market): share PR #1880 metric formatters with workbench module tables
Upstream PR #1880 landed number formatting (2-decimal values, count/amount/
percent/high-low coercion) for the legacy structured card while this branch
was in review. The workbench module cards render the same payload values, so
apply the same treatment for cross-path consistency: relocate the formatters
to marketReviewModuleMatching.ts (the pure-function module both components
may import — component files must not export non-components) and use them in
the module index table (close) and the breadth grid (counts, amount).
